Move Picker and BottomSheet to the @wordpress/components package
