--- packages/block-editor/src/components/mobile/bottom-sheet/button.native.js
@@ -1,32 +1,0 @@
-/**
- * External dependencies
- */
-import { TouchableOpacity, View, Text } from 'react-native';
-
-/**
- * Internal dependencies
- */
-import styles from './styles.scss';
-
-export default function Button( props ) {
-	const {
-		onPress,
-		disabled,
-		text,
-		color,
-	} = props;
-
-	return (
-		<TouchableOpacity
-			accessible={ true }
-			onPress={ onPress }
-			disabled={ disabled }
-		>
-			<View style={ { flexDirection: 'row', justifyContent: 'center' } }>
-				<Text style={ { ...styles.buttonText, color } }>
-					{ text }
-				</Text>
-			</View>
-		</TouchableOpacity>
-	);
-}
