+++ packages/components/src/mobile/bottom-sheet/cellStyles.android.scss
