+++ packages/components/src/mobile/bottom-sheet/cellStyles.ios.scss
