+++ packages/components/src/mobile/bottom-sheet/keyboard-avoiding-view.native.js
