+++ packages/format-library/src/link/modal.native.js
@@ -9,9 +9,9 @@ import { Platform } from 'react-native';
  */
 import { __ } from '@wordpress/i18n';
 import { Component } from '@wordpress/element';
-import { BottomSheet } from '@wordpress/block-editor';
 import { prependHTTP } from '@wordpress/url';
 import {
+	BottomSheet,
 	withSpokenMessages,
 } from '@wordpress/components';
 import {
